pyproject.toml
[tool.poetry]
name = "openai-summarization-app"
version = "0.1.0"
description = ""
authors = ["DrBenedictPorkins <flybaby6p2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.26.0"
langchain = "^0.0.285"
openai = "^0.28.0"
python-dotenv = "^1.0.0"
tiktoken = "^0.4.0"
tenacity = "^8.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
